build: Launch session allocator as current user, always tunnels with www-data
Compute session tunnels must be built with the same user as nginx,
so that it can forward requests to the compute sessions. There is,
though, no reason why the server and compute sessions
can't run with any other user.

This patch makes it so that the development server is run with the
same user that is running the Taskfile.

The actual .service in the .deb package stays with www-data,
which seems like an appropriate user for running the daemon.