Fixes recenter and ViewerDriver.snapTo
Opening images of different resolutions causes neuroglancer to change its
idea of what viewer.navigationState.pose.position means; it is measured in
"units of the finest open resolution". This would break the recenter button
when images of different resolutons were open, since the snapTo was always
setting position as if it was measured in voxels of the curret dataset on
display.